feat(radio)!: remove top level subcomponents
