+++ src/components/Radio/Radio.tsx
@@ -2,13 +2,11 @@ import clsx from 'clsx';
 import React from 'react';
 import { useId } from '../../util/useId';
 import type { EitherInclusive } from '../../util/utility-types';
-import type { RadioInputProps } from '../RadioInput';
-import RadioInput from '../RadioInput';
-import type { RadioLabelProps } from '../RadioLabel';
-import RadioLabel from '../RadioLabel';
+import Icon from '../Icon';
+import { InputLabel, type InputLabelProps } from '../InputLabel/InputLabel';
 import styles from './Radio.module.css';
 
-export type RadioProps = RadioInputProps & {
+type RadioProps = RadioInputProps & {
   /**
    * HTML id attribute. If not passed, this component
    * will generate an id to use for accessibility.
@@ -17,7 +15,7 @@ export type RadioProps = RadioInputProps & {
   /**
    * Size of the radio label.
    */
-  size?: RadioLabelProps['size'];
+  size?: InputLabelProps['size'];
 } & EitherInclusive<
     {
       /**
@@ -32,6 +30,86 @@ export type RadioProps = RadioInputProps & {
       'aria-label': string;
     }
   >;
+type RadioInputProps = Omit<
+  React.InputHTMLAttributes<HTMLInputElement>,
+  'id' | 'size'
+> & {
+  /**
+   * Additional classnames passed in for styling.
+   */
+  className?: string;
+  /**
+   * Radio ID. Used to connect the input with a label for accessibility purposes.
+   */
+  id?: string;
+};
+
+const RadioSvg = ({
+  checked,
+  disabled,
+}: {
+  checked?: boolean;
+  disabled?: boolean;
+}) => {
+  const iconClassName = clsx(
+    styles['radio__icon'],
+    disabled && styles['radio__icon--disabled'],
+  );
+  return (
+    <Icon
+      className={iconClassName}
+      name={checked ? 'radio-selected' : 'radio-unselected'}
+      purpose="decorative"
+      size="1.625rem"
+    />
+  );
+};
+
+/**
+ * Radio input element, exported for greater flexibility.
+ * You must provide an `id` prop and connect it to a visible label.
+ */
+export const RadioInput = ({
+  checked = false,
+  className,
+  disabled,
+  ...other
+}: RadioInputProps) => {
+  return (
+    <span
+      className={clsx(
+        styles['input__wrapper'],
+        disabled && styles['input__wrapper--disabled'],
+      )}
+    >
+      <input
+        checked={checked}
+        className={clsx(className, styles['radio__input'])}
+        disabled={disabled}
+        type="radio"
+        {...other}
+      />
+      <RadioSvg checked={checked} disabled={disabled} />
+    </span>
+  );
+};
+
+/**
+ * Radio label element, styles and relies on the InputLabel component.
+ */
+export const RadioLabel = ({
+  className,
+  size = 'lg',
+  ...other
+}: InputLabelProps) => {
+  const componentClassName = clsx(
+    styles['radio__label'],
+    styles[`radio__label--${size}`],
+    className,
+  );
+
+  return <InputLabel className={componentClassName} size={size} {...other} />;
+};
 
 /**
  * `import {Radio} from "@chanzuckerberg/eds";`
