--- src/components/RadioInput/RadioInput.module.css
@@ -1,70 +1,0 @@
-/*------------------------------------*\
-    # RADIO INPUT
-\*------------------------------------*/
-
-/**
- * Wraps the visually hidden radio input element and the visible sibling svg element.
- */
-.input__wrapper {
-  position: relative;
-  /* Centers the radio icon in the wrapper. */
-  display: inline-flex;
-  align-items: center;
-  /* Aligns the radio with the first line of the label. */
-  align-self: flex-start;
-}
-/**
- * The visually hidden input element for the radio. The visual radio is provided by an svg element.
- */
-.radio__input {
-  /* Removes default margins placed by browser for radioes. */
-  margin: 0;
-  /* Remove the radio from the page flow, positioning it on top of the SVG. */
-  position: absolute;
-  left: 0.25rem;
-  /* Set same dimensions as the RadioSvg element. */
-  height: var(--eds-size-2);
-  width: var(--eds-size-2);
-  /**
-   * Hide the input element visually.
-   * This ensures the radio is still "physically" present so that all users,
-   * especially on touch screen readers, still interact with the real radio element
-   * where it would naturally be present.
-   */
-  opacity: 0;
-}
-
-/**
- * The visible radio svg icon element.
- */
-.radio__icon {
-  color: var(--eds-theme-color-radio-brand-background);
-  /* Creates space for the border so that there's no jitter when the focus border is visible. */
-  border: var(--eds-border-width-sm) solid transparent;
-}
-.radio__input:focus-visible + .radio__icon {
-  border: var(--eds-border-width-sm) solid var(--eds-theme-color-focus-ring);
-  border-radius: var(--eds-border-radius-full);
-}
-
-@supports not selector(:focus-visible) {
-  .radio__input:focus + .radio__icon {
-    border: var(--eds-border-width-sm) solid var(--eds-theme-color-focus-ring);
-    border-radius: var(--eds-border-radius-full);
-  }
-}
-
-/**
- * The disabled variant of the visible radio svg icon.
- */
-.radio__icon--disabled {
-  color: var(--eds-theme-color-icon-disabled);
-}
-
-/**
- * The disabled status of the visually hidden input element.
- */
-.radio__input:disabled {
-  /* Needed since the input element overlays the visible svg icon for user input and cursor. */
-  cursor: not-allowed;
-}
