--- src/components/RadioInput/index.ts
@@ -1,2 +1,0 @@
-export { RadioInput as default } from './RadioInput';
-export type { RadioInputProps } from './RadioInput';
