--- src/components/RadioLabel/RadioLabel.module.css
@@ -1,21 +1,0 @@
-/*------------------------------------*\
-    # RADIO LABEL
-\*------------------------------------*/
-
-/**
- * Text that labels a radio input.
- */
-.radio__label {
-  position: relative;
-}
-
-/**
- * Radio label size variants.
- * Used for centering the label with the radio input button.
- */
-.radio__label--md {
-  top: 0.125rem;
-}
-.radio__label--lg {
-  top: 0.0625rem;
-}
