--- src/components/RadioLabel/index.ts
@@ -1,2 +1,0 @@
-export { RadioLabel as default } from './RadioLabel';
-export type { RadioLabelProps } from './RadioLabel';
